Change npm command from prepublishOnly to prepare
Right now, if one tries to add a Github fork of the project as a dependency, the project cannot properly be used, because dist/ is not committed to the repo; it is ignored by Git.

This commit changes the "prepublishOnly" transition script to "prepare", meaning "npm run build" will be run not only before publishing the package to npm, but also when installing the package. This means, dist/ will now automatically be built when if this project is added as a dependency if not already built.

The alternative to be to commit dist/ to the repo, but that is inelegant.